Pull orientationbasedvalue into its own file
